pyproject.toml
[project]
name = "shallow_learning_llm"
version = "0.0.1"
description = ""
authors = [
    {name = "songbinghui", email = "soongbinghui@gmail.com"},
]
dependencies = [
    "jupyterlab>=4.0.8",
    "numpy>=1.25.2",
    "matplotlib>=3.8.1",
    "python-dotenv>=1.0.0",
    "openai>=1.3.6",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
up = "jupyter lab"